Esqueletos das funcs de query (E os queries a serem usados), alguns bugfixes

+++ projeto/src/main/java/com/bd/ufrn/projeto/services/UserService.java
@@ -54,7 +54,7 @@ public class UserService {
         return newUser;
     }
 
-    public User findUser(Long id) {
+    public User findUser(Integer id) {
         return userRepository.findById(id);
     }
